Adapt to change to make all heights size_t and row.height uint32_t.

+++ include/bitcoin/server/utility/fetch_helpers.hpp
@@ -37,7 +37,7 @@ static BC_CONSTEXPR size_t point_size = hash_size + sizeof(uint32_t);
 // fetch_history stuff
 
 bool BCS_API unwrap_fetch_history_args(wallet::payment_address& address,
-    uint32_t& from_height, const message& request);
+    size_t& from_height, const message& request);
 
 void BCS_API send_history_result(const code& ec,
     const chain::history_compact::list& history, const message& request,
